Now all errors will be displayed only in debug mode
I decided to hide all the errors when DebugLevel set to 0.
Turns out, Excel will throw popups on every damn error and stop the script. For example, if script tries to change protected sheet, Excel won't put an error in Immediate window, it will throw everything to the user.
Basically, it's the same thing as it was with backgrounds and shapes.

Why Excel does this? No one knows.
Why didn't I know or even think that there are so many things that can go wrong? Guess, I'm stupid! :D