Added compatibility info for Office 2021
Partially compatible with Office 2021 (as of build 14228.20204). Dummies from Microsoft swapped light text theme colors with dark colors which resulted in Excel itself being incompatible with files created in some (not all) earlier versions. They might fix this behavior later. Since it's Excel bug, this Add-In won't attempt to fix it because it'll introduce even more inconsistent behavior.

Are they doing it on purpose?!